requirements.txt
numpy==1.20.1
PyAudio==0.2.11
scipy==1.6.1

Flask==1.1.2
PyMySQL==1.0.2
python-dotenv==0.15.0
